Moved relevant functions and tests into package
